requirements.txt
mcp>=1.0.0
torch
torchvision
torchaudio
transformers
accelerate
qwen-vl-utils
Pillow
